backend/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.0.6
django-allauth==0.50.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-spectacular==0.23.1
idna==3.3
inflection==0.5.1
jsonschema==4.9.0
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.11
